requirements.txt
zorg>=0.0.1
pillow>=3.0.0
pytesseract==0.1.6
